requirements.txt
aiogram==3.12.0
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.1
uvloop>=0.19.0; sys_platform != 'win32'
